LargeModels/FineTuning/group-7/AIRAG/requirements.txt
# 核心框架
langchain>=0.2.0,<0.4.0
langchain-community>=0.2.0,<0.4.0
langchain-openai>=0.1.0,<0.3.0
langchain-huggingface>=0.0.3

# 向量数据库
faiss-cpu>=1.9.0
chromadb>=0.4.0

# 嵌入模型
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.0.0

# 文档处理
pypdf>=4.0.0
python-docx>=1.0.0
openpyxl>=3.1.0

# 检索器
rank-bm25>=0.2.0

# 评测工具
ragas>=0.1.0

# Web UI
gradio>=4.0.0

# 数据分析和可视化
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0

# 工具库
python-dotenv>=1.0.0
tqdm>=4.65.0
loguru>=0.7.0
pydantic>=2.0.0
rich>=13.0.0

